@@ -1,12 +1,12 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import App from "./components/app/App";
 
 import "./style/style.scss";
 
-ReactDOM.render(
+ReactDOM.createRoot(document.getElementById("root"))
+.render(
   <React.StrictMode>
     <App />
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
