@@ -1,9 +1,11 @@
-import './comicsList.scss';
+import {useState, useEffect} from "react";
+import { Link } from "react-router-dom";
 
-import {useState, useEffect, useRef} from "react";
 import useMarvelService from '../../services/MarvelService';
 import Spinner from "../spinner/spinner";
 import ErrorMessage from "../errorMessage/ErrorMessage";
+
+import './comicsList.scss';
 
 
 
@@ -45,11 +47,11 @@ const ComicsList = (props) => {
         const items = arr.map((item, i) => {
           return (
             <li className="comics__item" key={i}>
-                <a href="#">
+                <Link to={`/comics/${item.id}`}>
                     <img src={item.thumbnail} alt={item.title} className="comics__item-img"/>
                     <div className="comics__item-name">{item.title}</div>
                     <div className="comics__item-price">{item.price}</div>
-                </a>
+                </Link>
             </li>
           );
         });
